RCLL-250 move logic from controller to courtcaseservice
