+++ server/controllers/recall/checkSentencesController.ts
@@ -10,6 +10,8 @@ import {
   isManualCaseSelection,
 } from '../../helpers/formWizardHelper'
 import ManageOffencesService from '../../services/manageOffencesService'
+import CourtCaseService from '../../services/CourtCaseService'
+import { services } from '../../services'
 
 export default class CheckSentencesController extends RecallBaseController {
   middlewareSetup() {
@@ -32,24 +34,29 @@ export default class CheckSentencesController extends RecallBaseController {
     return super.locals(req, res)
   }
 
-  async getOffenceNameTitle(req: FormWizard.Request, offenceCodes: string[]) {
-    return new ManageOffencesService().getOffenceMap(offenceCodes, req.user.token)
-  }
+  // async getOffenceNameTitle(req: FormWizard.Request, offenceCodes: string[]) {
+  //   return new ManageOffencesService().getOffenceMap(offenceCodes, req.user.token)
+  // }
+
 
   async loadOffenceNames(req: FormWizard.Request, res: Response, next: () => void) {
-    try {
-      const summarisedSentenceGroups = getSummarisedSentenceGroups(req)
-      const offenceCodes = summarisedSentenceGroups
-        .flatMap(group => group.sentences || [])
-        .map(charge => charge.offenceCode)
-        .filter(code => code)
+    /// service.getOffenceNameMap()
 
-      if (offenceCodes.length > 0) {
-        const offenceNameMap = await this.getOffenceNameTitle(req, offenceCodes)
-        res.locals.offenceNameMap = offenceNameMap
-      } else {
-        res.locals.offenceNameMap = {}
-      }
+    try {
+      // const courtCaseService = new CourtCaseService());
+      res.locals.offenceNameMap = req.services.courtCaseService.getOffenceNameMap(req)
+      // const summarisedSentenceGroups = getSummarisedSentenceGroups(req)
+      // const offenceCodes = summarisedSentenceGroups
+      //   .flatMap(group => group.sentences || [])
+      //   .map(charge => charge.offenceCode)
+      //   .filter(code => code)
+
+      // if (offenceCodes.length > 0) {
+      //   const offenceNameMap = await this.getOffenceNameTitle(req, offenceCodes)
+      //   res.locals.offenceNameMap = offenceNameMap
+      // } else {
+      //   res.locals.offenceNameMap = {}
+      // }
       next()
     } catch (error) {
       // eslint-disable-next-line no-console
